move signon-a-computer component into the shared module and rename device to DeviceType to DeviceTypeService
