+++ VideoWeb/VideoWeb/ClientApp/src/app/app.module.ts
@@ -17,7 +17,6 @@ import { OnTheDayModule } from './on-the-day/on-the-day.module';
 import { HomeComponent } from './home/home.component';
 import { WaitingSpaceModule } from './waiting-space/waiting-space.module';
 import { SendVideoEventsComponent } from './send-video-events/send-video-events.component';
-import { DeviceDetectorModule } from 'ngx-device-detector';
 
 export function getSettings(configService: ConfigService) {
   return () => configService.loadConfig();
@@ -37,8 +36,7 @@ export function getSettings(configService: ConfigService) {
     SecurityModule,
     WaitingSpaceModule,
     OnTheDayModule,
-    AppRoutingModule,
-    DeviceDetectorModule.forRoot(),
+    AppRoutingModule
   ],
   providers: [
     { provide: APP_INITIALIZER, useFactory: getSettings, deps: [ConfigService], multi: true },
