+++ VideoWeb/VideoWeb/ClientApp/src/app/on-the-day/on-the-day.module.ts
@@ -15,7 +15,7 @@ import { ParticipantHearingsComponent } from './participant-hearings/participant
 import { SelfTestComponent } from './self-test/self-test.component';
 import { SwitchOnCameraMicrophoneComponent } from './switch-on-camera-microphone/switch-on-camera-microphone.component';
 import { VideoCheckComponent } from './video-check/video-check.component';
-import { SignonAComputerComponent } from './signon-a-computer/signon-a-computer.component';
+import { SignonAComputerComponent } from '../shared/signon-a-computer/signon-a-computer.component';
 
 @NgModule({
   imports: [
